+++ packages/console-types/tap-snapshots/src/types/config.test.ts.test.cjs
@@ -5034,7 +5034,7 @@ Object {
                                 "type": "boolean",
                               },
                               "secretKey": Object {
-                                "pattern": "^[a-zA-Z0-9-_.]*$",
+                                "pattern": "^((\\\\{\\\\{([A-Z])([A-Z0-9_]*)\\\\}\\\\})|[a-zA-Z0-9-_.]*)$",
                                 "type": "string",
                                 "x-validation-error-id": "serviceSecretKey.patternError",
                               },
@@ -5770,7 +5770,7 @@ Object {
                           "type": "boolean",
                         },
                         "secretKey": Object {
-                          "pattern": "^[a-zA-Z0-9-_.]*$",
+                          "pattern": "^((\\\\{\\\\{([A-Z])([A-Z0-9_]*)\\\\}\\\\})|[a-zA-Z0-9-_.]*)$",
                           "type": "string",
                           "x-validation-error-id": "serviceSecretKey.patternError",
                         },