+++ packages/console-sdk-microfrontend/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @mia-platform/console-sdk-microfrontend
 
+## 0.2.40
+
+### Patch Changes
+
+- Updated dependencies [5316390]
+  - @mia-platform/console-types@0.25.0
+
+## 0.2.39
+
+### Patch Changes
+
+- 9fe0e15: feat(events): Add event type "REFRESH_EXTENSIONS"
+
 ## 0.2.38
 
 ### Patch Changes