+++ packages/console-sdk-vite-helpers/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @mia-platform/console-sdk-vite-helpers
 
+## 0.0.80
+
+### Patch Changes
+
+- @mia-platform/console-sdk-microfrontend@0.2.40
+
+## 0.0.79
+
+### Patch Changes
+
+- Updated dependencies [9fe0e15]
+  - @mia-platform/console-sdk-microfrontend@0.2.39
+
 ## 0.0.78
 
 ### Patch Changes